PSB Pathfinders_Aman Mathur/PSB Pathfinders_Aman Mathur/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
alembic
email-validator
bcrypt==4.0.1

pymysql
cryptography

python-dotenv
